// client/src/models/product.ts
export type Product = {
  id: number;                  // Identificador único del producto
  name: string;                // Nombre del producto
  description: string;         // Descripción detallada del producto
  price: number;               // Precio del producto
  pictureUrl: string;          // URL de la imagen del producto
  type: string;                // Tipo o categoría del producto
  brand: string;               // Marca del producto
  quantityInStock: number;     // Cantidad disponible en inventario
};